+++ internal/providers/volcasr/asr_errors_test.go
@@ -11,8 +11,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/coder/websocket"
 	"github.com/evjohn-icu/timingdex/internal/providers/common"
-	"nhooyr.io/websocket"
 )
 
 func TestASRErrorPaths(t *testing.T) {