+++ internal/providers/volcasr/asr_test.go
@@ -13,8 +13,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/coder/websocket"
 	"github.com/evjohn-icu/timingdex/internal/providers/common"
-	"nhooyr.io/websocket"
 )
 
 func TestDecodeFrameRejectsOversizedPayload(t *testing.T) {